refactor : cache module 분리
