+++ api/build.gradle.kts
@@ -2,6 +2,7 @@
 dependencies {
     implementation(project(":shared:common"))
     implementation(project(":shared:bizppurio-support"))
+    implementation(project(":shared:cache-support"))
 
     implementation ("org.springframework.boot:spring-boot-starter-validation")
     implementation ("org.springframework.boot:spring-boot-starter-data-jpa")
