+++ settings.gradle.kts
@@ -4,3 +4,4 @@ include(":shared:discord-support")
 include(":shared:common")
 include(":api")
 include("shared:bizppurio-support")
+include("shared:cache-support")
