+++ shared/bizppurio-support/build.gradle.kts
@@ -1,8 +1,9 @@
 
 dependencies {
     implementation(project(":shared:common"))
+    implementation(project(":shared:cache-support"))
+
     implementation ("org.springframework.boot:spring-boot-starter-webflux")
-    implementation ("org.springframework.boot:spring-boot-starter-web")
     implementation("org.springframework:spring-tx")
 }
