+++ shared/bizppurio-support/src/main/java/com/yedu/bizppurio/support/application/usecase/BizppurioCheckStep.java
@@ -5,9 +5,9 @@ import com.fasterxml.jackson.databind.ObjectMapper;
 import com.yedu.bizppurio.support.application.constant.BizppurioResponseCode;
 import com.yedu.bizppurio.support.application.dto.req.MessageStatusRequest;
 import com.yedu.bizppurio.support.event.publisher.BizppurioModuleEventPublisher;
+import com.yedu.cache.support.RedisRepository;
 import com.yedu.common.event.bizppurio.MatchingConfirmTeacherEvent.IntroduceWriteFinishTalkEvent;
 import com.yedu.common.event.bizppurio.MatchingParentsEvent.ParentsClassNoticeEvent;
-import com.yedu.common.redis.RedisRepository;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Component;
