+++ shared/cache-support/build.gradle.kts
@@ -1,5 +1,6 @@
 
 dependencies {
+    implementation ("org.springframework.boot:spring-boot-starter-data-redis")
 }
